@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~>4.0.0'
+gem 'rails', '~>4.1.0'
 gem 'activerecord-session_store'
 
 gem 'select2-rails'
