@@ -48,10 +48,7 @@ class Task < ActiveRecord::Base
     task = nil
     begin
       Task.transaction do
-        task = Task.find(:first, 
-                         :order => "created_at", 
-                         :conditions => {:status=> Task::STATUS_INQUEUE}, 
-                         :lock => true)
+        task = Task.where(status: Task::STATUS_INQUEUE).where(lock: true).first
         if task!=nil
           task.status = status
           task.save!
